chg: organise range and geo code into validators modules
